utf8_heavy: Skip unnecessary operations
The mktables generated tables are well-formed, already sorted, and with
no extra items such as "+utf8::foo".  Thus we don't have to do these
operations on them, but they are required on user-defined properties,
and should $list be passed in as a parameter.

This patch moves the code that does this to just the user-defined area