Add newXS_len_flags
It accepts a length as well as a pv for the name.

Since newXS_flags is marked with M in embed.fnc and is undocumented,
technically policy allows me to change it, but there are files
throughout cpan/ that use newXS_flags.  So it seemed safer to add a
new function.