Make constant sub redef warnings obey scope
In perldiag, this is listed as (S), which means that outside of any
use/no warnings scope it always warns, regardless of $^W.

But this warning was ignoring use/no warnings, too.

There were actually tests for this oddity, but I think those were
added by mistake, or this was just not thought through.  I cannot see
how this is not a bug.