Put sub redef warnings in one place
The logic surrounding subroutine redefinition warnings (to warn or not
to warn?) was in three places.  Over time, they drifted apart, to thepoint that newXS was following completely different rules.  It was
only warning for redefinition of functions in the autouse namespace.
Recent commits have brought it into conformity with the other redefi-
nition warnings.

Obviously it’s about time we put it in one function.