utf8_heavy.pl: Add inversion status to cache key
Contrary to what the debug statement said, what is being returned is a
swash, and that swash is different from one that comes from the same
file but differs in inversion, and so changing the INVERT_IT element
messes things up for any existing one.  Heretofore it hasn't mattered
because the swash returned is always a copy, and so it actually hasn't
created any problems.  But future commits will stop the copying, so this
would create problems then.

The file will now have to be re-'do'ne to get an inverted list from it.