In roffitall, explicitly use perl to run rofftoc.
Previously roffitall invoked rofftoc as ./rofftoc (implying a shell script),
and in turn rofftoc used the "sh or perl" construction to cause sh to pass
itself to perl. This was a bit obfuscated at the best of times, and broke
completely when commit ff906f87ce151941 removed execute permission in 2010.