Remove duplicate comment from sv.h
The ‘or lexical’ in this ‘glob or lexical is just a copy’ comment was
added in perl 5.001 (748a9306), which added closures.

This comment was later duplicated in commit 120ff8e9d (‘Document a
sixth use for SVf_FAKE’), but in a clearer fashion.

This commit removes the ‘or lexical’ part, as it is now confusing.