Add Term::ReadLine to write_buildcustomize.pl.
If the build fails while building extensions, it's nice to have
the debugger available to help figure out what went wrong.  You
couldn't do that before because lib/perl5db.pl depends on
Term::ReadLine, which wouldn't be available since it hadn't been
built yet.  This commit makes Term::ReadLine available via the
same mechanism that makes other libraries available to miniperl
during the build.

An alternative would be to remove the debugger's dependency on
Term::ReadLine, but that would be more work and more risk for a
situation that hopefully doesn't come up that often.